mybooks is driving me insane rn

+++ cmput301w19t15/app/src/main/java/com/example/cmput301w19t15/Activity/Profile.java
@@ -82,8 +82,9 @@ public class Profile extends AppCompatActivity {
         inputName.setText(user.getName());
         inputEmail.setText(user.getEmail());
         inputPhoneNumber.setText(user.getPhone());
+
         try {
-            rating.setRating(user.getRating().intValue());
+            rating.setRating(user.getRating());
         }
         catch (Exception e){
             Log.d("testing","no user rating set");
